+++ test/integrationTest/testAssets/testRoot/main.c
@@ -8,5 +8,8 @@ int main()
     printf("Release Mode");
 #endif
 
+    int i = 2;
+    int j = i * 3;
+
     return 0;
 }